Add @inline to some functions (#65)
This is dangerous and we should instead find the actual reason for the
inconsistency. I'll leave this as a draft for now and will look into it
more later. Since `Base.@assume_effects` requires Julia 1.8 this will
fail unit tests anyways.

Having these functions infer `:consistent` correctly by the compiler can
significantly help with optimization in the compiler, so eventually we
should try to make sure that most of QEDcore functionality has as many
effects inferred as possible.

Edit: I have removed the assumptions again because they are not safe to
make (at least not for all element types). We should come back to this
at a later point. For now, I have only added some `@inbounds`
statements.